@@ -1,55 +1,28 @@
 {
   "name": "nfc-travel-vcard",
-  "version": "1.0.0",
   "private": true,
+  "version": "0.1.0",
   "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "start:compat": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
-    "build:compat": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "dayjs": "^1.11.13",
-    "express": "^5.1.0",
-    "mssql": "^11.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-helmet-async": "^2.0.5",
-    "react-router": "^6.16.0",
-    "react-router-dom": "^6.16.0",
-    "react-scripts": "4.0.3",
-    "serve": "^14.2.4",
-    "swiper": "^6.8.4",
-    "typescript": "^4.1.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/express": "^5.0.2",
-    "@types/mssql": "^9.1.7",
-    "@types/node": "^22.15.24",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-helmet": "^6.1.11"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5"
   }
 }
